Added a temporary REST API to Gaffer.  This is a bit of Python code which implements a web server, and then calls out through py4j to get Gaffer to execute things.
This is temporary.  Gaffer has most of the support for a REST interface in
development, but I'm too impatient, so created this.  It implements a subset
of the API which will be implemented by Gaffer, so no porting effort required
to move to the real API when it is finished.